requirement.txt
pyecharts==1.9.0
joblib==0.17.0
jieba==0.42.1
pypinyin==0.41.0
python-levenshtein==0.12.2